added menu, and basic reply functionality

+++ src/main/java/tetris/Tetris.java
@@ -7,6 +7,8 @@ import java.awt.Rectangle;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
+
+import javax.swing.JButton;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.Timer;
@@ -14,7 +16,7 @@ import javax.swing.Timer;
 
 public class Tetris implements ActionListener{
 
-    static void o(String s) {
+    static public void o(String s) {
         System.out.println(s);
     }
 
@@ -34,6 +36,14 @@ public class Tetris implements ActionListener{
     static Input input;
     static long score;
     static JLabel label;
+    static Mongo db;
+    static public boolean replaying;
+    static JButton play;
+    static JButton highScore;
+    static JButton replay;
+    static final int BUTTON_W = 150;
+    static final int BUTTON_H = 50;
+
     
     Tetris() {    
         input = new Input();
@@ -46,15 +56,65 @@ public class Tetris implements ActionListener{
         frame.setLocationRelativeTo(null);
         frame.setVisible(true); 
         placed = new ArrayList<>();
-        
+        db = new Mongo();
         label = new JLabel("SCORE: " + score);
         label.setBounds(10, 10, 100, 30);
         label.setFont(new Font("Arial", Font.BOLD, 18));
         window.add(label);
+        play = new JButton();
+        replay = new JButton();
+        highScore = new JButton();
+        play.setText("PLAY");
+        replay.setText("REPLAY");
+        highScore.setText("HIGH SCORE");
+        play.setSize(BUTTON_W, BUTTON_H);
+        replay.setSize(BUTTON_W, BUTTON_H);
+        highScore.setSize(BUTTON_W, BUTTON_H);
+        play.setLocation(W / 2 - BUTTON_W / 2, 100);
+        replay.setLocation(W / 2 - BUTTON_W / 2, 300);
+        highScore.setLocation(W / 2 - BUTTON_W / 2, 200);
+        play.setBackground(Color.PINK);
+        replay.setBackground(Color.PINK);
+        highScore.setBackground(Color.PINK);
+        play.addActionListener(this);
+        replay.addActionListener(this);
+        highScore.addActionListener(this);
+        play.setActionCommand("play");
+        replay.setActionCommand("replay");
+        highScore.setActionCommand("highscore");
+        window.add(play);
+        window.add(replay);
+        window.add(highScore);
+        hideMenu();
     }
     
     @Override
     public void actionPerformed(ActionEvent e) {
+        String action = e.getActionCommand();
+        if (action == null) {
+            action = "";
+        }
+        if (action.equals("play")) {
+            hideMenu();
+            startGame(false);
+            return;
+        } else if (action.equals("replay")) {
+            hideMenu();
+            startGame(true);
+            return;
+        } else if (action.equals("highscore")) {
+            return;
+        }
+        if (replaying) {
+            Thingy newCurr = db.getThingy();
+            if (newCurr == null) {
+                stopGame();
+            }
+            if (curr.id != newCurr.id) {
+                placed.add(curr);
+            }
+            spawn(newCurr);
+        }
         ++ticks;
         if (input.right) {
             curr.right();
@@ -75,18 +135,43 @@ public class Tetris implements ActionListener{
         tempRefreshSum += refreshTime;
         if (tempRefreshSum >= delay) {
             tempRefreshSum = 0;
-            if (dropAuto) {                
+            if (dropAuto && !replaying) {                
                 curr.drop();        
             }
             dropAuto = true;
-        }     
+        } 
+        increaseSpeed();
+        removeLevels();        
+        if (!replaying) {
+            db.addThingy(curr);
+        }
         if (!curr.dropped) {
             placed.add(curr);
-            spawn();
-        }        
-        increaseSpeed();
-        removeLevels();
+            spawn(null);
+        }    
         window.repaint();
+    }
+
+    static void showMenu() {
+        play.setEnabled(true);
+        play.setVisible(true);
+        if (db.storesGame()) {
+            replay.setEnabled(true);
+        }        
+        replay.setVisible(true);
+        highScore.setEnabled(true);
+        highScore.setVisible(true);
+        window.setBackground(Color.WHITE);
+    }
+
+    static void hideMenu() {
+        play.setEnabled(false);
+        play.setVisible(false);
+        replay.setEnabled(false);
+        replay.setVisible(false);
+        highScore.setEnabled(false);
+        highScore.setVisible(false);
+        window.setBackground(Color.CYAN);
     }
     
     void removeLevel(int y) {
@@ -130,9 +215,13 @@ public class Tetris implements ActionListener{
             }
         }
     }
-    
-    void spawn() {
-        curr = Generator.getThingy();        
+
+    void spawn(Thingy thingy) {        
+        if (thingy == null) {
+            curr = Generator.getThingy();  
+        } else {
+            curr = thingy;
+        }
         for (Rectangle r : curr.recs) {
             for (Thingy t : placed) {
                 if (t.collide(r)) {
@@ -143,22 +232,38 @@ public class Tetris implements ActionListener{
         }
     } 
     
-    void startGame() {
+    void startGame(boolean repl) {        
+        hideMenu();
         Input.clear();
+        placed.clear();
+        score = 0;
+        label.setText("SCORE: " + score);
+        curr = shadow = null;
         delay = 1000;
         ticks = 0;
         tempRefreshSum = 0;
         dropAuto = true;
         score = 0;
-        spawn();
+        if (repl) {
+            replay();
+        } else {
+            db.clear();
+            spawn(null);
+        }
         timer = new Timer(refreshTime, this);
         timer.start(); 
         o("start");
     }
-    
+
+
+    boolean ok = true;
+
     void stopGame() {
         o("stop");
+        replaying = false;
         timer.stop();
+        db.saveGame();        
+        showMenu();
     }
     
     void increaseSpeed() {
@@ -182,8 +287,13 @@ public class Tetris implements ActionListener{
         curr.draw(g);     
     }
     
+    void replay() {
+        replaying = true;
+        curr = db.getThingy();
+    }
+
     public static void main(String[] args) {
         game = new Tetris();
-        game.startGame();
+        showMenu();
     }
 }
